+++ lib/components/appbar_controller.dart
@@ -7,6 +7,14 @@ enum AppbarPropagation {
   next,
 }
 
+/// This class provides essential functionality and roles for managing and tracking the appbar.
+/// 
+/// See Also, It performs the following roles:
+/// 
+/// - Manages the competition for consuming scroll offsets among app bars.
+/// - Defines and manages the state of app bars.
+/// - Rebuilds the layout through alignment and listeners.
+/// - Or other related tasks.
 class AppBarController extends Listenable {
   /// The values defines the instance of [AppBarPosition].
   final _positions = <AppBarPosition>[];
@@ -28,17 +36,21 @@ class AppBarController extends Listenable {
     notifyListeners();
   }
 
-  attach(AppBarPosition position) {
+  /// Delegates the task of adding the appbar position to this controller
+  /// to ensure it can be reliably detached and disposed later.
+  void attach(AppBarPosition position) {
     assert(!_positions.contains(position), "Already attached in this controller.");
     _positions.add(position);
   }
 
-  detach(AppBarPosition position) {
+  /// Delegates the task of detaching and disposing of the appbar position
+  /// to ensure consistency with [attach] function.
+  void detach(AppBarPosition position) {
     assert(_positions.contains(position), "Already not attached in this controller.");
     _positions.remove(position);
   }
 
-  /// Returns attached a appbar position in this controller by given index.
+  /// Returns attached the appbar position in this controller by given index.
   AppBarPosition at(int index) {
     if (_positions.length < index && index < 0) {
       throw FlutterError("The given index overflowed attached appbar positions length.");
@@ -78,6 +90,14 @@ class AppBarController extends Listenable {
     for (final it in _positions) { it.align(position); }
   }
 
+  /// Delegates all positions and context from a given controller to itself
+  /// and removes all positions from the given controller, ensuring that
+  /// each appbar position exists only once in the controller.
+  void delegateFrom(AppBarController other) {
+    _positions.clear();
+    _positions.addAll(other._positions..clear());
+  }
+
   @override
   void addListener(VoidCallback listener) {
     assert(!_listeners.contains(listener), "Already exists a given listener.");
@@ -90,9 +110,15 @@ class AppBarController extends Listenable {
     _listeners.remove(listener);
   }
 
-  notifyListeners() {
+  /// Notifies that the state related to the controller has changed.
+  void notifyListeners() {
     for (final listener in _listeners) {
       listener.call();
     }
+  }
+
+  /// Disposes all instances related the controller(e.g. [TouchRippleEffect]).
+  void dispose() {
+    _positions.toList().forEach((position) => detach(position));
   }
 }