+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 1.0.0-dev1
+- 😀 Initial publishing.
 
-* TODO: Describe initial release.
+## 1.0.0-dev2
+- Proceed with documentation of README.md and its overall source-code by adding comments.
+- Renamed from `AppBarAlign` to `AppBarAlignmentCommand` and from `AppBarAlignBehavior` to `AppBarAlignmentBehavior`.
+- Fixed an issue by implementing code that delegates all appbar states and operations from the existing appbar controller to the new app bar controller.